fix keyboards back steps

+++ auto_bot/handlers/order/handlers.py
@@ -97,9 +97,9 @@ def payment_personal_order(update, context):
     query.edit_message_text(edit_text)
     payload = f"{chat_id} {query.message.message_id} {price}"
     payment_request(chat_id,
-                    chat_id,
                     payload,
                     price)
+
 
 def back_step_to_finish_personal(update, context):
     query = update.callback_query
@@ -165,11 +165,10 @@ def from_address(update, context):
 
 
 def to_the_address(update, context):
-
     query = update.callback_query
     chat_id = update.effective_chat.id
-    state = int(redis_instance().hget(str(chat_id), 'state'))
-    if state == FROM_ADDRESS:
+    state = redis_instance().hget(str(chat_id), 'state')
+    if state and state == str(FROM_ADDRESS):
         buttons = [[InlineKeyboardButton(f'{NOT_CORRECT_ADDRESS}', callback_data='From_address 0')], ]
         address = update.message.text
         addresses = buttons_addresses(address)
@@ -198,8 +197,8 @@ def to_the_address(update, context):
 def payment_method(update, context):
     query = update.callback_query
     chat_id = update.effective_chat.id
-    state = int(redis_instance().hget(str(chat_id), 'state'))
-    if state == TO_THE_ADDRESS:
+    state = redis_instance().hget(str(chat_id), 'state')
+    if state and state == str(TO_THE_ADDRESS):
         address = update.message.text
         buttons = [[InlineKeyboardButton(f'{NOT_CORRECT_ADDRESS}', callback_data='To_the_address 0')], ]
         addresses = buttons_addresses(address)
@@ -216,8 +215,9 @@ def payment_method(update, context):
             context.bot.send_message(chat_id=chat_id, text=wrong_address_request)
             to_the_address(update, context)
     else:
+        back_step = "Wrong_place" if redis_instance().hget(str(chat_id), 'personal_flag') else "Right_place"
         query.edit_message_text(add_info_text)
-        query.edit_message_reply_markup(inline_add_info_kb())
+        query.edit_message_reply_markup(inline_add_info_kb(back_step))
         redis_instance().hdel(str(chat_id), 'state')
 
 
@@ -280,7 +280,7 @@ def first_address_check(update, context):
         from_address(update, context)
 
 
-def payment_request(chat_id_client, start_parameter, payload, price: int):
+def payment_request(chat_id_client, payload, price: int):
     prices = [LabeledPrice(label=payment_price, amount=int(price) * 100)]
 
     # Sending a request for payment
@@ -290,7 +290,6 @@ def payment_request(chat_id_client, start_parameter, payload, price: int):
                      payload=payload,
                      provider_token=os.environ["PAYMENT_TOKEN"],
                      currency=payment_currency,
-                     start_parameter=start_parameter,
                      prices=prices,
                      photo_url=os.environ["BOT_URL_IMAGE_TAXI"],
                      need_shipping_address=False,
@@ -350,7 +349,6 @@ def order_create(update, context):
         redis_instance().hmset(f"{order_data['chat_id_client']}_", order_data)
         payment_request(order_data['chat_id_client'],
                         order_data['chat_id_client'],
-                        order_data['chat_id_client'],
                         order_data['sum'])
     else:
         order_create_task.delay(order_data)
@@ -385,13 +383,13 @@ def increase_order_price(update, context):
 
 def choose_date_order(update, context):
     query = update.callback_query
+    chat_id = update.effective_chat.id
+    duty = UserBank.get_duty(chat_id)
     if query.data == "Personal_order":
         redis_instance().hset(str(update.effective_chat.id), 'personal_flag', query.data)
         query.edit_message_text(order_date_text)
-        query.edit_message_reply_markup(inline_choose_date_kb())
-    chat_id = update.effective_chat.id
-    duty = UserBank.get_duty(chat_id)
-    if duty.duty >= int(ParkSettings.get_value('USER_DUTY')):
+        query.edit_message_reply_markup(inline_choose_date_kb("Personal_driver"))
+    elif duty.duty >= int(ParkSettings.get_value('USER_DUTY')):
         query.edit_message_text(duty_of_user)
         query.edit_message_reply_markup(user_duty())
     else:
@@ -401,7 +399,7 @@ def choose_date_order(update, context):
             query.edit_message_text(order_not_payment)
         else:
             query.edit_message_text(order_date_text)
-            query.edit_message_reply_markup(inline_choose_date_kb())
+            query.edit_message_reply_markup(inline_choose_date_kb("Back_to_main"))
 
 
 def time_order(update, context):
@@ -472,7 +470,7 @@ def client_reject_order(update, context):
         else:
             portmone = Portmone()
             total_amount = report.total_amount
-            total_sum = total_amount - (total_amount * 0.02)
+            total_sum = total_amount * 0.98
             if total_sum > cancel:
                 portmone.return_amount(total_sum-cancel, report.provider_payment_charge_id, return_money)
                 report_for_client = f'{return_money_from_system} {get_money} {cancel}{report.currency}'
@@ -669,7 +667,6 @@ def handle_order(update, context):
             total = order.sum - total_amount
             if total > 0:
                 payment_request(order.chat_id_client,
-                                order.pk,
                                 f'{order.pk} {query.message.message_id}',
                                 total)
             else:
@@ -685,7 +682,6 @@ def handle_order(update, context):
 
         else:                                       # first cash second card
             payment_request(order.chat_id_client,
-                            order.pk,
                             f'{order.pk} {query.message.message_id}',
                             order.sum)
 
@@ -694,7 +690,6 @@ def payment_duty(update, context):
     chat_id = str(update.effective_chat.id)
     duty = UserBank.get_duty(chat_id)
     payment_request(chat_id,
-                    chat_id,
                     chat_id,
                     duty.duty)
     redis_instance().hset(chat_id, 'duty', 1)
@@ -737,34 +732,34 @@ def successful_payment(update, context):
         order = Order.objects.filter(chat_id_client=chat_id, status_order=Order.CANCELED,
                                      payment_method=price_inline_buttons[4].split()[1]).last()
         create_report_payment(successful_payment, order)
-    elif order.type_order == Order.STANDARD_TYPE:
-        fleet_order(order)
-        context.bot.send_message(chat_id=order.driver.chat_id, text=trip_paymented)
-        text_to_client(order, complete_order_text, button=inline_comment_for_client())
-    elif order.type_order == Order.PERSONAL_TYPE:
-        create_report_payment(successful_payment, order)
-        update_hours = int(redis_instance().hget(chat_id, 'hours'))
-        order.update(
-            payment_hours=F('payment_hours') + update_hours,
-            sum=F('sum') + int(list_payload[2])
-        )
-        context.bot.delete_message(chat_id=chat_id, message_id=redis_instance().hget(chat_id, 'invoice'))
-        context.bot.send_message(chat_id=chat_id,
-                                 text=update_hours_text(update_hours))
-        try:
-            msg = redis_instance().hget(str(order.first().driver.chat_id), "driver_msg")
-            context.bot.delete_message(chat_id=order.first().driver.chat_id, message_id=msg)
-        except BadRequest as e:
-            get_logger().error(e)
-        driver_msg = context.bot.send_message(chat_id=order.first().driver.chat_id,
-                                              text=update_hours_driver_text(update_hours))
-        redis_instance().hset(str(order.first().driver.chat_id), "driver_msg", driver_msg.message_id)
-        create_report_payment(successful_payment, order)
+    if order:
+        if order.type_order == Order.STANDARD_TYPE:
+            fleet_order(order)
+            context.bot.send_message(chat_id=order.driver.chat_id, text=trip_paymented)
+            text_to_client(order, complete_order_text, button=inline_comment_for_client())
+        elif order.type_order == Order.PERSONAL_TYPE:
+            create_report_payment(successful_payment, order)
+            update_hours = int(redis_instance().hget(chat_id, 'hours'))
+            order.update(
+                payment_hours=F('payment_hours') + update_hours,
+                sum=F('sum') + int(list_payload[2])
+            )
+            context.bot.delete_message(chat_id=chat_id, message_id=redis_instance().hget(chat_id, 'invoice'))
+            context.bot.send_message(chat_id=chat_id,
+                                     text=update_hours_text(update_hours))
+            try:
+                msg = redis_instance().hget(str(order.first().driver.chat_id), "driver_msg")
+                context.bot.delete_message(chat_id=order.first().driver.chat_id, message_id=msg)
+            except BadRequest as e:
+                get_logger().error(e)
+            driver_msg = context.bot.send_message(chat_id=order.first().driver.chat_id,
+                                                  text=update_hours_driver_text(update_hours))
+            redis_instance().hset(str(order.first().driver.chat_id), "driver_msg", driver_msg.message_id)
+            create_report_payment(successful_payment, order)
     elif personal_order:
         save_location_to_redis(chat_id)
         user_data = redis_instance().hgetall(chat_id)
         client = Client.get_by_chat_id(chat_id)
-
         data = {'order_time': datetime.fromisoformat(user_data['time_order']),
                 'chat_id_client': chat_id,
                 'phone_number': client.phone_number,
@@ -779,7 +774,7 @@ def successful_payment(update, context):
         if user_data.get('info'):
             data['info'] = user_data['info']
         order = Order.objects.create(**data)
-        context.bot.delete_message(chat_id=chat_id, message_id=redis_instance().hget(chat_id, 'invoice'))
+        # context.bot.delete_message(chat_id=chat_id, message_id=redis_instance().hget(chat_id, 'invoice'))
         message = bot.send_message(chat_id=chat_id,
                                    text=client_personal_info(order),
                                    reply_markup=inline_reject_order(order.pk))
